pyproject.toml
[tool.poetry]
name = "budgetapi"
version = "0.1.0"
description = "An API to track spendings."
authors = ["Danila Kritsky <danilakritsky@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
mypy = "^0.990"
bandit = "^1.7.4"
safety = "^2.3.1"


[tool.poetry.group.api.dependencies]
django = "^4.1.3"
djangorestframework = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
preview = true
line-length = 79

[tool.mypy]
show_error_codes = true
allow_redefinition = true
ignore_missing_imports = true
disable_error_code = ["union-attr", "return", "no-redef"]

[tool.isort]
profile = "black"
line_length = 79
lines_after_imports = 2